still won't reliably end moves
When the last node is clicked, it still won't always trip the doMove()
properly. I believe that it is throwing errors, but I don't know how to
check these error on code blocks. NOTE: For some reason game.cpp got
updated on mine as well from lines 360 to 364 and 417 to 420. It appears
to be in the doMove() put I don't remember editing anything.